Rename header step counter
