+++ GUI/src/components/HeaderStepCounter/HeaderStep/HeaderStep.scss
