+++ GUI/src/components/HeaderStepCounter/HeaderStepCounter.scss
